+++ system-tests/maven-external-upload-test/auto-resolve-revision-project/src/main/java/org/example/SUT.java
@@ -1,0 +1,7 @@
+package org.example;
+
+public class SUT {
+    public void hello() {
+        System.out.println("hello");
+    }
+}